Add files via upload
Objective: Our client is an E-commerce website owner who is not happy with the performance 
on his website in terms of profit despite having good engagement on the website. 

 Outcome: We are predicting the order status and analyzing the percentage of the orders which 
are getting cancelled or completed. With the products clustering we can give our client the list 
of products he can concentrate more on, so we are giving him inventory management solutions 
and customer segmentation based on the customers behavior on the website. 

Techniques Performed: Data preprocessing techniques, EDA, Various models were built and a 
comparative analysis was done on the performance of each model based on the Accuracy score
from which one best model that had better performance was chosen.

Final Model: RandomForest classifier was chosen as final model with accuracy score of 92.04% 
on train and test set.